Step G2: register primitives in the source-gen JSON context; normalize enums
ConstantNode.PrimitiveValue is declared object?, so System.Text.Json resolves
its JsonTypeInfo by each value's runtime type. Under source generation every
such type must be registered or serialization throws NotSupportedException.
The context listed the 19 node types and no primitives at all.

Registered the exact set ExpressionToNodeTranslator.IsPrimitive admits.

Registration alone was not sufficient: IsPrimitive also admits enums, whose
runtime type is the concrete enum and therefore can never be pre-registered.
Enums are now converted to their underlying integral value at capture --
which ConstantNode's own doc comment already anticipated ("enum-underlying") --
and rebuilt from the node's TypeNode by CoercePrimitive.

Used explicit registration rather than chaining a reflection-based fallback
resolver: a fallback would resolve everything at runtime and silently defeat
the AOT/trimming goal (requirements 4.5, ADR-008 constraint 8).

Also fixed a latent bug in the enum rebuild path now being relied on:
Enum.ToObject(underlying, element.GetInt32()) throws for enums with long or
ulong backing. Uses GetInt64.

207 -> 289 passing, 206 -> 124 failing of 413.

+++ src/InfoCarrier.Core/Expressions/NodeToExpressionTranslator.cs
@@ -90,7 +90,9 @@ public class NodeToExpressionTranslator
             Type underlying = Nullable.GetUnderlyingType(targetType) ?? targetType;
             if (underlying.IsEnum)
             {
-                return Enum.ToObject(underlying, element.GetInt32());
+                // GetInt64 (not GetInt32) so enums with long/ulong backing survive; the
+                // client sends the underlying integral value, never the enum name.
+                return Enum.ToObject(underlying, element.GetInt64());
             }
 
             if (underlying == typeof(string)) return element.GetString();
